requirements.txt
alembic==1.10.2
Babel==2.12.1
blinker==1.5
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
dnspython==2.3.0
dominate==2.7.0
email-validator==1.3.1
Flask==2.2.3
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==1.0.0
pytz==2022.7.1
SQLAlchemy==2.0.6
typing_extensions==4.5.0
visitor==0.1.3
Werkzeug==2.2.3
WTForms==3.0.1
